# swagger.yaml
openapi: 3.0.0
info:
  title: learn express and swagger
  description: learn express and swagger writing docs
  version: 1.0.1
  contact:
    email: hanshirlukku@gmail.com
    url: "https://hanshirlukku.dev"

# servers:
#   - url: "https://localhost:4000/api/v1"
#     description: localhost secure
#   - url: "https://localhost:4000/api/v1"
#     description: localhost regular

servers:
  - url: "{protocol}://localhost:4000/api/{version}"
    description: localhost secure
    variables:
      version:
        enum:
          - v1
          - v2
        default: v1
      protocol:
        enum:
          - http
          - https
        default: http

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    BearerAuth:
      type: http
      scheme: bearer

paths:
  /lco:
    get:
      tags:
        - String
      summary: returns a greet message from LCO
      responses:
        200:
          description: All good and sucess
          content:
            application/json:
              schema:
                type: string
                example: "My string"
        400:
          description: Bad request
        500:
          description: Internel server error

  /lcoobject:
    get:
      tags:
        - Object
      summary: returns a unique course
      responses:
        200:
          description: All good and sucess
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: number
        400:
          description: Bad request
        500:
          description: Internel server error

  /courses:
    get:
      tags:
        - Array
      summary: returns all course
      responses:
        200:
          description: All good and sucess
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    price:
                      type: number
        400:
          description: Bad request
        500:
          description: Internel server error

  /mycourse/{courseId}:
    get:
      tags:
        - String
      summary: returns course based on request id
      parameters:
        - name: courseId
          in: path
          required: true
          default: 22
          schema:
            type: string
      responses:
        200:
          description: All good and sucess
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: number
        400:
          description: Bad request
        500:
          description: Internel server error

  /addCourse:
    post:
      tags:
        - String
      summary: add a new course to exisisting courses
      consumes:
        - application/json
      produces:
        - application/json
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                price:
                  type: number
      responses:
        200:
          description: All good and sucess
          content:
            application/json:
              schema:
                type: boolean
        400:
          description: Bad request
        500:
          description: Internel server error

  /coursequery:
    get:
      tags:
        - String
      summary: trying to learn about query
      parameters:
        - name: location
          in: query
          required: true
          schema:
            type: string
            enum: [delhi, landon, jaipur]
        - name: device
          in: query
          required: true
          schema:
            type: string
            enum: [web, mobile]
      responses:
        200:
          description: All good and sucess
          content:
            application/json:
              schema:
                type: object
                properties:
                  location:
                    type: string
                  device:
                    type: string
        400:
          description: Bad request
        500:
          description: Internel server error

  /courseupload:
    post:
      tags:
        - String
      summary: Uploading course image
      parameters:
        - in: header
          name: auth
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: All good and sucess
        400:
          description: Bad request
        500:
          description: Internel server error